orders/src/events/listener/__test__/expiration-completed-listener.test.ts
import { OrderExpirationEvent, orderStatus } from "@akticket/common";
import mongoose from "mongoose";
import { Message } from "node-nats-streaming";
import { Order } from "../../../models/orders";
import { Ticket } from "../../../models/tickets";
import { natsWrapper } from "../../../nats-wrapper";
import { OrderExpirationListener } from "../expiration-completed-listener";

const setup = async () => {
  const listener = new OrderExpirationListener(natsWrapper.client);

  const ticket = Ticket.build({
    id: new mongoose.Types.ObjectId().toHexString(),
    title: "concert",
    price: 15,
  });
  await ticket.save();

  const order = Order.build({
    userId: "asd",
    expiresAt: new Date(),
    status: orderStatus.Created,
    ticket,
  });
  await order.save();

  const data: OrderExpirationEvent["data"] = {
    orderId: order.id,
  };

  //@ts-ignore
  const msg: Message = {
    ack: jest.fn(),
  };

  return { listener, order, ticket, data, msg };
};

it("Updates the order status to cancel", async () => {
  const { listener, order, data, msg } = await setup();

  await listener.onMessage(data, msg);

  const updatedOrder = await Order.findById(order.id);

  expect(updatedOrder?.status).toEqual(orderStatus.Cancelled);
});

it("Sends an Order cancelled event", async () => {
  const { listener, order, data, msg } = await setup();

  await listener.onMessage(data, msg);

  expect(natsWrapper.client.publish).toHaveBeenCalled();
});

it("acks the message", async () => {
  const { listener, order, data, msg } = await setup();

  await listener.onMessage(data, msg);

  expect(msg.ack).toHaveBeenCalled();
});
